AV-214603 Moving HTTPRoute Validation to Graph layer
