+++ ako-gateway-api/status/gatewayclass_status.go
@@ -73,36 +73,36 @@ func (o *gatewayClass) BulkUpdate(key string, options []status.StatusOptions) {
 	// TODO: Add this code when we publish the status from the rest layer
 }
 
-func (o *gatewayClass) Patch(key string, obj runtime.Object, status *status.Status, retryNum ...int) {
+func (o *gatewayClass) Patch(key string, obj runtime.Object, status *status.Status, retryNum ...int) error {
 	retry := 0
 	if len(retryNum) > 0 {
 		retry = retryNum[0]
 		if retry >= 5 {
 			utils.AviLog.Errorf("key: %s, msg: Patch retried 5 times, aborting", key)
-			return obj, errors.New("Patch retried 5 times, aborting")
+			return errors.New("Patch retried 5 times, aborting")
 		}
 	}
 
 	gatewayClass := obj.(*gatewayv1.GatewayClass)
 	if o.isStatusEqual(&gatewayClass.Status, status.GatewayClassStatus) {
-		return obj, nil
+		return nil
 	}
 
 	patchPayload, _ := json.Marshal(map[string]interface{}{
 		"status": status.GatewayClassStatus,
 	})
-	updatedObject, err := akogatewayapilib.AKOControlConfig().GatewayAPIClientset().GatewayV1().GatewayClasses().Patch(context.TODO(), gatewayClass.Name, types.MergePatchType, patchPayload, metav1.PatchOptions{}, "status")
+	_, err := akogatewayapilib.AKOControlConfig().GatewayAPIClientset().GatewayV1().GatewayClasses().Patch(context.TODO(), gatewayClass.Name, types.MergePatchType, patchPayload, metav1.PatchOptions{}, "status")
 	if err != nil {
 		utils.AviLog.Warnf("key: %s, msg: there was an error in updating the GatewayClass status. err: %+v, retry: %d", key, err, retry)
 		updatedObj, err := akogatewayapilib.AKOControlConfig().GatewayApiInformers().GatewayClassInformer.Lister().Get(gatewayClass.Name)
 		if err != nil {
 			utils.AviLog.Warnf("GatewayClass not found %v", err)
-			return updatedObj, err
+			return err
 		}
 		return o.Patch(key, updatedObj, status, retry+1)
 	}
 	utils.AviLog.Infof("key: %s, msg: Successfully updated the GatewayClass %s status %+v %v", key, gatewayClass.Name, utils.Stringify(status), err)
-	return updatedObject, nil
+	return nil
 }
 
 func (o *gatewayClass) isStatusEqual(old, new *gatewayv1.GatewayClassStatus) bool {
