+++ ako-gateway-api/status/httproute_status.go
@@ -84,37 +84,37 @@ func (o *httproute) BulkUpdate(key string, options []status.StatusOptions) {
 	// TODO: Add this code when we publish the status from the rest layer
 }
 
-func (o *httproute) Patch(key string, obj runtime.Object, status *status.Status, retryNum ...int) {
+func (o *httproute) Patch(key string, obj runtime.Object, status *status.Status, retryNum ...int) error {
 	retry := 0
 	if len(retryNum) > 0 {
 		retry = retryNum[0]
 		if retry >= 5 {
 			utils.AviLog.Errorf("key: %s, msg: Patch retried 5 times, aborting", key)
-			return obj, errors.New("Patch retried 5 times, aborting")
+			return errors.New("Patch retried 5 times, aborting")
 		}
 	}
 
 	httpRoute := obj.(*gatewayv1.HTTPRoute)
 	if o.isStatusEqual(&httpRoute.Status, status.HTTPRouteStatus) {
-		return obj, nil
+		return nil
 	}
 
 	patchPayload, _ := json.Marshal(map[string]interface{}{
 		"status": status.HTTPRouteStatus,
 	})
-	updatedObject, err := akogatewayapilib.AKOControlConfig().GatewayAPIClientset().GatewayV1().HTTPRoutes(httpRoute.Namespace).Patch(context.TODO(), httpRoute.Name, types.MergePatchType, patchPayload, metav1.PatchOptions{}, "status")
+	_, err := akogatewayapilib.AKOControlConfig().GatewayAPIClientset().GatewayV1().HTTPRoutes(httpRoute.Namespace).Patch(context.TODO(), httpRoute.Name, types.MergePatchType, patchPayload, metav1.PatchOptions{}, "status")
 	if err != nil {
 		utils.AviLog.Warnf("key: %s, msg: there was an error in updating the HTTPRoute status. err: %+v, retry: %d", key, err, retry)
 		updatedObj, err := akogatewayapilib.AKOControlConfig().GatewayApiInformers().HTTPRouteInformer.Lister().HTTPRoutes(httpRoute.Namespace).Get(httpRoute.Name)
 		if err != nil {
 			utils.AviLog.Warnf("HTTPRoute not found %v", err)
-			return updatedObj, err
+			return err
 		}
 		return o.Patch(key, updatedObj, status, retry+1)
 	}
 
 	utils.AviLog.Infof("key: %s, msg: Successfully updated the HTTPRoute %s/%s status %+v", key, httpRoute.Namespace, httpRoute.Name, utils.Stringify(status))
-	return updatedObject, nil
+	return nil
 }
 
 func (o *httproute) isStatusEqual(old, new *gatewayv1.HTTPRouteStatus) bool {
